requirements.txt
jproperties
python_telegram_bot==20.0a4